+++ WhistleWindLobotomyMod/Patches/ExtendedSniperLogicPatches.cs
@@ -1,17 +1,19 @@
 ﻿using DiskCardGame;
 using HarmonyLib;
 using InscryptionAPI.Card;
+using InscryptionAPI.Helpers.Extensions;
 using InscryptionCommunityPatch.Card;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using WhistleWind.AbnormalSigils;
+using WhistleWind.AbnormalSigils.StatusEffects;
 using WhistleWind.Core.Helpers;
-using WhistleWindLobotomyMod;
 using WhistleWindLobotomyMod.Core;
 using static InscryptionCommunityPatch.Card.SniperFix;
 
 // custom version of the Sniper fix that adds Marksman compatibility and the special behaviour of Blue Star and Judgement Bird
-namespace WhistleWind.AbnormalSigils.Patches
+namespace WhistleWindLobotomyMod.Patches
 {
     [HarmonyPatch]
     internal class ExtendedSniperLogicPatches
@@ -28,16 +30,6 @@ namespace WhistleWind.AbnormalSigils.Patches
 
             return true;
         }
-        [HarmonyPrefix, HarmonyPatch(typeof(SniperFix), nameof(SniperFix.SniperAttack))]
-        private static bool OverrideSniperSequence(CombatPhaseManager instance, CardSlot slot, ref IEnumerator __result)
-        {
-            if (IsJudgementBird(slot) || IsEnraged(slot) || IsRaging(slot))
-            {
-                __result = WstlSniperSequence(instance, slot);
-                return false;
-            }
-            return true;
-        }
         [HarmonyPostfix, HarmonyPatch(typeof(SniperFix), nameof(SniperFix.WillDieFromSharp))]
         private static void AddExtraChecks(CardSlot slot, ref bool __result)
         {
@@ -45,145 +37,110 @@ namespace WhistleWind.AbnormalSigils.Patches
             if (IsJudgementBird(slot))
                 __result = false;
         }
-        private static IEnumerator WstlSniperSequence(CombatPhaseManager instance, CardSlot slot)
-        {
-            List<CardSlot> opposingSlots = new(); // the slots to attacks
-            Singleton<ViewManager>.Instance.SwitchToView(Singleton<BoardManager>.Instance.CombatView, false, false);
-            Singleton<ViewManager>.Instance.Controller.LockState = ViewLockState.Locked;
-            int numAttacks = GetAttackCount(slot.Card);
-            Singleton<ViewManager>.Instance.Controller.SwitchToControlMode(Singleton<BoardManager>.Instance.ChoosingSlotViewMode, false);
-            Singleton<ViewManager>.Instance.Controller.LockState = ViewLockState.Unlocked;
-            Part1SniperVisualizer visualizer = null;
-            if (SaveManager.SaveFile.IsPart1)
-                visualizer = instance.GetComponent<Part1SniperVisualizer>() ?? instance.gameObject.AddComponent<Part1SniperVisualizer>();
 
-            if (IsRaging(slot))
+        [HarmonyPostfix, HarmonyPatch(typeof(SniperFix), nameof(SniperFix.DoSniperLogic))]
+        private static IEnumerator CustomSniperLogic(
+            IEnumerator enumerator, CombatPhaseManager instance, Part1SniperVisualizer visualizer,
+            List<CardSlot> opposingSlots, CardSlot slot, int numAttacks)
+        {
+            if (IsRaging(slot)) // if raging, target random slots
             {
-                opposingSlots = slot.Card.GetOpposingSlots();
+                opposingSlots.AddRange(slot.Card.GetOpposingSlots());
                 for (int i = 0; i < opposingSlots.Count; i++)
                 {
                     instance.VisualizeConfirmSniperAbility(opposingSlots[i]);
                     visualizer?.VisualizeConfirmSniperAbility(opposingSlots[i]);
                 }
+                yield break;
             }
-            else
+            if (IsEnraged(slot)) // if enraged, target the rival if they exist
             {
-                CardSlot enragedTarget = null;
-                if (IsEnraged(slot))
+                string nameToFind = "wstl_redHoodedMercenary";
+
+                if (slot.Card.Info.name == nameToFind)
+                    nameToFind = "wstl_willBeBadWolf";
+
+                CardSlot enragedTarget = BoardManager.Instance.AllSlotsCopy.Find(x => x.Card?.Info.name == nameToFind);
+                if (enragedTarget != null)
                 {
-                    string nameToFind;
-                    if (slot.Card.Info.name == "wstl_redHoodedMercenary")
-                        nameToFind = "wstl_willBeBadWolf";
-                    else
-                        nameToFind = "wstl_redHoodedMercenary";
+                    for (int i = 0; i < numAttacks; i++)
+                        opposingSlots.Add(enragedTarget);
 
-                    enragedTarget = BoardManager.Instance.AllSlotsCopy.Find(x => x.Card?.Info.name == nameToFind);
-                    if (enragedTarget != null)
+                    instance.VisualizeConfirmSniperAbility(enragedTarget);
+                    visualizer?.VisualizeConfirmSniperAbility(enragedTarget);
+                    yield break;
+                }
+            }
+
+            yield return enumerator;
+        }
+        [HarmonyPostfix, HarmonyPatch(typeof(SniperFix), nameof(SniperFix.DoAttackTargetSlotsLogic))]
+        private static IEnumerator CustomOpposingSlotsLogic(IEnumerator enumerator, CardSlot slot, CardSlot opposingSlot)
+        {
+            if (!IsJudgementBird(slot) || ImmuneToHanging(opposingSlot) || slot.Card.AttackIsBlocked(opposingSlot))
+            {
+                yield return enumerator;
+                yield break;
+            }
+
+            Singleton<ViewManager>.Instance.SwitchToView(Singleton<BoardManager>.Instance.CombatView);
+            opposingSlot.Card.FlipFaceDown(false);
+            opposingSlot.Card.Anim.StrongNegationEffect();
+            yield return new WaitForSeconds(0.4f);
+            yield return Execution(opposingSlot.Card);
+        }
+
+        [HarmonyPostfix, HarmonyPatch(typeof(SniperFix), nameof(SniperFix.PlayerTargetSelectedCallback))]
+        private static void CustomTargetCallback(CardSlot targetSlot, CardSlot attackingSlot)
+        {
+            if (IsJudgementBird(attackingSlot) && !ImmuneToHanging(targetSlot))
+                targetSlot.Card.AddTemporaryMod(new() { singletonId = executeId });
+        }
+        [HarmonyPostfix, HarmonyPatch(typeof(SniperFix), nameof(SniperFix.PlayerSlotCursorEnterCallback))]
+        private static void CustomCursorEnterCallback(CardSlot targetSlot)
+        {
+            InteractionCursor.Instance.ForceCursorType(ImmuneToHanging(targetSlot) ? CursorType.Target : CursorType.Sacrifice);
+        }
+        [HarmonyPostfix, HarmonyPatch(typeof(SniperFix), nameof(SniperFix.OpponentSelectTargetSlot))]
+        private static void CustomSelectTargetSlot(ref CardSlot __result, List<CardSlot> opposingSlots, List<PlayableCard> playerCards,
+            CardSlot attackingSlot, int numAttacks)
+        {
+            bool anyCards = playerCards.Count > 0;
+            if (anyCards && attackingSlot.Card != null)
+            {
+                int sinCount = attackingSlot.Card.GetStatusEffectStacks<Sin>();
+                if (sinCount > 0)
+                {
+                    List<CardSlot> filteredSlots = new()
+                        {
+                            GetStrongestKillableCard(anyCards, playerCards, opposingSlots, attackingSlot, numAttacks)?.Slot,
+                            GetFirstStrongestAttackableCard(anyCards, playerCards, opposingSlots, attackingSlot, numAttacks)?.Slot,
+                            GetFirstStrongestAttackableCardNoPreferences(anyCards, playerCards, opposingSlots, attackingSlot, numAttacks)?.Slot,
+                            __result
+                        };
+                    // remove null slots and sort by lowest sin count
+                    filteredSlots.RemoveAll(x => x == null);
+                    filteredSlots.Sort((a, b) => (a.Card?.GetStatusEffectStacks<Sin>() ?? 0) - (b.Card?.GetStatusEffectStacks<Sin>() ?? 0));
+
+                    if (filteredSlots.Exists(x => x.Card != null))
                     {
-                        for (int i = 0; i < numAttacks; i++)
-                            opposingSlots.Add(enragedTarget);
-
-                        instance.VisualizeConfirmSniperAbility(enragedTarget);
-                        visualizer?.VisualizeConfirmSniperAbility(enragedTarget);
+                        if (sinCount >= 5) // if we have 5+ Sin, target the card with the lowest sin count
+                            __result = filteredSlots[0];
+                        else
+                            __result = filteredSlots.Find(x => x.Card.GetStatusEffectStacks<Sin>() + sinCount >= 5) ?? __result;
                     }
                 }
-
-                if (enragedTarget == null)
-                {
-                    if (slot.Card.OpponentCard)
-                        yield return WstlOpponentLogic(instance, visualizer, opposingSlots, slot, numAttacks);
-                    else
-                        yield return WstlPlayerLogic(instance, visualizer, opposingSlots, slot, numAttacks);
-                }
             }
-
-            Singleton<ViewManager>.Instance.Controller.SwitchToControlMode(Singleton<BoardManager>.Instance.DefaultViewMode, false);
-            Singleton<ViewManager>.Instance.Controller.LockState = ViewLockState.Locked;
-            foreach (CardSlot opposingSlot in opposingSlots)
-            {
-                Singleton<ViewManager>.Instance.SwitchToView(Singleton<BoardManager>.Instance.CombatView);
-                if (IsJudgementBird(slot) && opposingSlot.Card != null && !ImmuneToHanging(opposingSlot) && !slot.Card.AttackIsBlocked(opposingSlot))
-                {
-                    opposingSlot.Card.FlipFaceDown(false);
-                    opposingSlot.Card.Anim.StrongNegationEffect();
-                    yield return new WaitForSeconds(0.4f);
-                    yield return Execution(opposingSlot.Card);
-                }
-                else
-                    yield return instance.SlotAttackSlot(slot, opposingSlot, opposingSlots.Count > 1 ? 0.1f : 0f);
-            }
-            instance.VisualizeClearSniperAbility();
-            visualizer?.VisualizeClearSniperAbility();
-            yield break;
         }
 
-        private static IEnumerator WstlPlayerLogic(CombatPhaseManager instance, Part1SniperVisualizer visualizer,
-        List<CardSlot> opposingSlots, CardSlot slot, int numAttacks)
+        private static bool ImmuneToHanging(CardSlot slot)
         {
-            for (int i = 0; i < numAttacks; i++)
+            if (slot.Card != null)
             {
-                instance.VisualizeStartSniperAbility(slot);
-                visualizer?.VisualizeStartSniperAbility(slot);
-                CardSlot cardSlot = Singleton<InteractionCursor>.Instance.CurrentInteractable as CardSlot;
-                if (cardSlot != null && opposingSlots.Contains(cardSlot))
-                {
-                    instance.VisualizeAimSniperAbility(slot, cardSlot);
-                    visualizer?.VisualizeAimSniperAbility(slot, cardSlot);
-                }
-                List<CardSlot> list = Singleton<BoardManager>.Instance.OpponentSlotsCopy;
-                yield return Singleton<BoardManager>.Instance.ChooseTarget(list, list,
-                    delegate (CardSlot s)
-                    {
-                        opposingSlots.Add(s);
-                        if (IsJudgementBird(slot) && !ImmuneToHanging(s))
-                            s.Card.AddTemporaryMod(new() { singletonId = executeId });
-
-                        instance.VisualizeConfirmSniperAbility(s);
-                        visualizer?.VisualizeConfirmSniperAbility(s);
-                    }, null, delegate (CardSlot s)
-                    {
-                        InteractionCursor.Instance.ForceCursorType(ImmuneToHanging(s) ? CursorType.Target : CursorType.Sacrifice);
-                        instance.VisualizeAimSniperAbility(slot, s);
-                        visualizer?.VisualizeAimSniperAbility(slot, s);
-                    }, () => false, CursorType.Target);
+                return slot.Card.HasAbility(Ability.MadeOfStone) || slot.Card.HasAnyOfTraits(Trait.Terrain, Trait.Pelt, AbnormalPlugin.ImmuneToInstaDeath);
             }
+            return true;
         }
-        private static IEnumerator WstlOpponentLogic(CombatPhaseManager instance, Part1SniperVisualizer visualizer,
-        List<CardSlot> opposingSlots, CardSlot slot, int numAttacks)
-        {
-            List<CardSlot> playerSlots = Singleton<BoardManager>.Instance.PlayerSlotsCopy;
-            List<PlayableCard> playerCards = playerSlots.FindAll(x => x.Card != null).ConvertAll((x) => x.Card);
-            bool anyCards = playerCards.Count > 0;
-
-            for (int i = 0; i < numAttacks; i++)
-            {
-                CardSlot attackSlot = slot.opposingSlot;
-                if (anyCards)
-                {
-                    PlayableCard strongestKillable = GetStrongestKillableCard(anyCards, playerCards, opposingSlots, slot, numAttacks);
-                    PlayableCard strongestAttackable = GetFirstStrongestAttackableCard(anyCards, playerCards, opposingSlots, slot, numAttacks);
-                    PlayableCard strongestAttackableNoPreferences = GetFirstStrongestAttackableCardNoPreferences(anyCards, playerCards, opposingSlots, slot, numAttacks);
-
-                    if (CanWin(opposingSlots, playerSlots, slot, numAttacks))
-                        attackSlot = GetFirstAvailableOpenSlot(opposingSlots, playerSlots, slot, numAttacks);
-                    else if (strongestKillable != null)
-                        attackSlot = strongestKillable.Slot;
-                    else if (strongestAttackable != null)
-                        attackSlot = strongestAttackable.Slot;
-                    else if (strongestAttackableNoPreferences != null)
-                        attackSlot = strongestAttackableNoPreferences.Slot;
-                }
-
-                opposingSlots.Add(attackSlot);
-                if (IsJudgementBird(slot) && !ImmuneToHanging(attackSlot))
-                    attackSlot.Card.AddTemporaryMod(new() { singletonId = executeId });
-
-                instance.VisualizeConfirmSniperAbility(attackSlot);
-                visualizer?.VisualizeConfirmSniperAbility(attackSlot);
-                yield return new WaitForSeconds(0.25f);
-            }
-        }
-        private static bool ImmuneToHanging(CardSlot slot) => slot.Card?.HasAnyOfTraits(Trait.Terrain, Trait.Pelt, AbnormalPlugin.ImmuneToInstaDeath) ?? true;
         private static bool IsJudgementBird(CardSlot slot) => slot?.Card != null && slot.Card.HasTrait(LobotomyCardManager.TraitExecutioner);
         private static bool IsEnraged(CardSlot slot) => slot?.Card != null && (slot.Card.GetComponent<CrimsonScar>()?.Enraged ?? false);
         private static bool IsRaging(CardSlot slot) => slot?.Card != null && slot.Card.HasSpecialAbility(BlindRage.specialAbility);